minor refactoring of root-finding functions + new introots function
